refactor into more composable responsibilties
