+++ app/workers/workflow_contents_dump_worker.rb
@@ -1,7 +1,9 @@
 require 'csv'
 
 class WorkflowContentsDumpWorker
+  include Sidekiq::Worker
   include DumpWorker
+  include DumpMailerWorker
   include RateLimitDumpWorker
 
   attr_reader :project
